Moved functions in CssGen that produce CSS common to all pages to a CommonCssGen.hs
